+++ client/src/App.jsx
@@ -4,7 +4,7 @@ import { Route, Routes, useLocation } from 'react-router-dom'
 import Loader from './common/Loader'
 import PageTitle from './components/PageTitle'
 import SignIn from './pages/Authentication/SignIn'
-import ECommerce from './pages/Dashboard/ECommerce'
+import ECommerce from './pages/Dashboard/DashboardHome'
 import Profile from './pages/Dashboard/Settings/Profile'
 import ProfessorIndex from './pages/Dashboard/Professor/ProfessorIndex'
 import RequireAuth from './features/auth/RequireAuth'
